shopfloor: cluster picking split scan tests
