+++ shopfloor/tests/__init__.py
@@ -12,7 +12,9 @@ from . import test_single_pack_transfer_putaway
 from . import test_cluster_picking_base
 from . import test_cluster_picking_batch
 from . import test_cluster_picking_select
-from . import test_cluster_picking_scan
+from . import test_cluster_picking_scan_line
+from . import test_cluster_picking_scan_destination
+from . import test_cluster_picking_is_zero
 from . import test_cluster_picking_skip
 from . import test_cluster_picking_stock_issue
 from . import test_cluster_picking_change_pack_lot
